requirements.txt
Flask
requests
flatten_dicts
Werkzeug==0.16.0
